+++ src/jsMain/kotlin/dev/inmo/jsuikit/utils/Percents.kt
@@ -1,0 +1,5 @@
+package dev.inmo.jsuikit.utils
+
+value class Percents(val int: Int) {
+    override fun toString(): String = "${int}%"
+}